fix sign up/log in page: deduplicate code and consolidate into one component
