+++ client/src/App.js
@@ -3,8 +3,7 @@ import './App.css';
 import { BrowserRouter, Switch, Route } from 'react-router-dom';
 import CssBaseline from '@material-ui/core/CssBaseline';
 import LandingPage from './components/LandingPage';
-import SignUp from './components/SignUp';
-import LogIn from './components/LogIn';
+import SignUpOrLogIn from './components/SignUpOrLogIn';
 import GuestUI from './components/GuestUI';
 
 function App() {
@@ -14,8 +13,8 @@ function App() {
             <BrowserRouter>
                 <Switch>
                     <Route exact path="/" component={LandingPage} />
-                    <Route exact path="/signup" component={SignUp} />
-                    <Route exact path="/login" component={LogIn} />
+                    <Route exact path="/signup" component={() => <SignUpOrLogIn isSigningUp={true} />} />
+                    <Route exact path="/login" component={() => <SignUpOrLogIn isSigningUp={false} />} />
                     <Route exact path="/guestui" component={GuestUI} />
                 </Switch>
             </BrowserRouter>
