+++ client/src/components/ChatMessageDisplay.js
@@ -1,5 +1,4 @@
 import React from 'react';
-import { makeStyles } from '@material-ui/core/styles';
 import Paper from '@material-ui/core/Paper';
 import ChatBoxMessage from './ChatBoxMessage';
