--- client/src/components/LogIn.js
@@ -1,66 +1,0 @@
-import React, { useState } from 'react';
-import { makeStyles } from '@material-ui/core/styles';
-import Input from '@material-ui/core/Input';
-import InputLabel from '@material-ui/core/InputLabel';
-import InputAdornment from '@material-ui/core/InputAdornment';
-import IconButton from '@material-ui/core/IconButton';
-import FormControl from '@material-ui/core/FormControl';
-import TextField from '@material-ui/core/TextField';
-import MailOutlineIcon from '@material-ui/icons/MailOutline';
-import Visibility from '@material-ui/icons/Visibility';
-import VisibilityOff from '@material-ui/icons/VisibilityOff';
-import SendIcon from '@material-ui/icons/Send';
-import Button from '@material-ui/core/Button';
-
-const useStyles = makeStyles((theme) => ({
-    margin: {
-        margin: theme.spacing(1),
-    },
-    button: {
-        marginTop: 20,
-    },
-}));
-
-function Login() {
-    const classes = useStyles();
-    let [showPassword, setShowPassword] = useState(true);
-
-    return (
-        <div class="App-header">
-            <FormControl className={classes.margin}>
-                <InputLabel htmlFor="input-with-icon-adornment">Email</InputLabel>
-                <Input
-                    id="input-with-icon-adornment"
-                    startAdornment={
-                        <InputAdornment position="start">
-                            <MailOutlineIcon />
-                        </InputAdornment>
-                    }
-                />
-            </FormControl>
-            <TextField
-                className={classes.margin}
-                id="input-with-icon-textfield"
-                label="Password"
-                type={showPassword ? 'password' : 'text'}
-                InputProps={{
-                    startAdornment: (
-                        <InputAdornment position="start">
-                            <IconButton
-                                aria-label="toggle password visibility"
-                                onClick={() => setShowPassword(!showPassword)}
-                            >
-                                {showPassword ? <Visibility /> : <VisibilityOff />}
-                            </IconButton>
-                        </InputAdornment>
-                    ),
-                }}
-            />
-            <Button variant="outlined" color="primary" className={classes.button} startIcon={<SendIcon />}>
-                Log in
-            </Button>
-        </div>
-    );
-}
-
-export default Login;
